requirements.txt
amqp==5.0.5
APScheduler==3.7.0
bidict==0.21.2
billiard==3.6.3.0
boto3==1.17.12
botocore==1.20.12
celery==5.0.5
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
coinpayments-py==0.1.0
cryptography==3.4.6
dnspython==2.1.0
ecdsa==0.16.1
email-validator==1.1.2
Flask==1.1.2
Flask-DotEnv==0.1.2
Flask-Login==0.5.0
Flask-MySQLdb==0.2.0
Flask-QRcode==3.0.0
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
gunicorn==20.0.4
idna==2.10
install==1.3.4
ipapi==1.0.4
itsdangerous==1.1.0
Jinja2==2.11.3
jmespath==0.10.0
kombu==5.0.2
MarkupSafe==1.1.1
mysqlclient==2.0.3
netifaces==0.10.6
paypal-checkout-serversdk==1.0.1
paypalhttp==1.0.0
Pillow==8.1.2
prompt-toolkit==3.0.16
pycparser==2.20
pyOpenSSL==20.0.1
python-dateutil==2.8.1
python-dotenv==0.15.0
python-engineio==4.0.0
pytz==2021.1
qrcode==6.1
redis==3.5.3
redis-server==6.0.9
requests==2.25.1
s3transfer==0.3.4
six==1.15.0
SQLAlchemy==1.3.23
stripe==2.56.0
tzlocal==2.1
urllib3==1.26.3
vine==5.0.0
wcwidth==0.2.5
Werkzeug==1.0.1
WTForms==2.3.3
